fix: move header and footer to layout
