+++ src/app/layout.tsx
@@ -3,6 +3,7 @@ import './globals.css';
 import { Inter } from 'next/font/google';
 import MUIThemeProvider from '@utils/MUIThemeProvider';
 import TanStackQueryProvider from '@utils/TanStackQueryProvider';
+import ClientSideLayout from '@components/sharedComponents/layout/ClientSideLayout';
 
 const inter = Inter({ subsets: ['latin'] });
 
@@ -16,9 +17,13 @@ const RootLayout: FC<PropsWithChildren> = ({ children }) => {
   return (
     <html lang="en">
       <body className={inter.className}>
-        <TanStackQueryProvider>
-          <MUIThemeProvider>{children}</MUIThemeProvider>
-        </TanStackQueryProvider>
+        <main>
+          <TanStackQueryProvider>
+            <MUIThemeProvider>
+              <ClientSideLayout>{children}</ClientSideLayout>
+            </MUIThemeProvider>
+          </TanStackQueryProvider>
+        </main>
       </body>
     </html>
   );
