+++ src/components/sharedComponents/layout/ClientSideLayout.tsx
@@ -1,0 +1,17 @@
+'use client';
+import React, { FC, PropsWithChildren } from 'react';
+import Header from './pageHeader/Header';
+import Footer from './pageFooter/Footer';
+import { Container } from '@mui/material';
+
+const ClientSideLayout: FC<PropsWithChildren> = ({ children }) => {
+  return (
+    <>
+      <Header />
+      <Container maxWidth="xl">{children}</Container>
+      <Footer />
+    </>
+  );
+};
+
+export default ClientSideLayout;
